Fix integer overflow when computing 'application' memory
Notice that the result of (inactive page count + active page count) *
page size can exceed 2^32, so we need to do the computation using
uint64.

+++ plugins/memory/freebsdbackend.cpp
@@ -70,7 +70,7 @@ void FreeBsdMemoryBackend::makeSensors()
 
 unsigned long long FreeBsdMemoryBackend::pagesToBytes(uint32_t pages)
 {
-    return m_pageSize * pages;
+    return m_pageSize * static_cast<unsigned long long>(pages);
 }
 
 void FreeBsdMemoryBackend::update()
